Feature/strapi 5

+++ packages/core/admin/screens/List/components/TableRow/index.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useState } from 'react';
+import React, { FC } from 'react';
 import {
   Typography,
   Box,
@@ -7,19 +7,16 @@ import {
   Flex,
   IconButton,
 } from '@strapi/design-system';
-import {
-  request,
-  useNotification,
-} from '@strapi/helper-plugin';
-import { Attribute, Entity } from '@strapi/strapi';
+import { useNotification, getFetchClient } from '@strapi/strapi/admin';
 import { useIntl } from 'react-intl';
 import { Trash, ExternalLink, Pencil } from '@strapi/icons';
-import { useHistory } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import DeleteConfirmModal from '../DeleteConfirmModal';
-import { Config } from '../../../../../server/admin-api/config';
+import { Config } from '../../../../../server/config';
+import { UrlAliasEntity } from '../../../../types/url-aliases';
 
 type Props = {
-  row: Attribute.GetValues<'plugin::webtools.url-alias'>;
+  row: UrlAliasEntity;
   checked?: boolean;
   onDelete?: () => void;
   updateValue: () => any;
@@ -35,28 +32,28 @@ const TableRow: FC<Props> = ({
   config,
   onDelete,
 }) => {
-  const toggleNotification = useNotification();
+  const { toggleNotification } = useNotification();
+  const { get, post } = getFetchClient();
   const { formatMessage } = useIntl();
-  const [openDeleteModal, setOpenDeleteModal] = useState(false);
-  const { push } = useHistory();
+  const navigate = useNavigate();
 
   const handleClick = (path: string) => {
-    request(`/webtools/url-alias/editLink?path=${path}`, { method: 'GET' })
-      .then((res: { link: string }) => {
-        push(res.link);
+    get<{ link: string }>(`/webtools/url-alias/editLink?path=${path}`)
+      .then((res) => {
+        navigate(res.data.link);
       })
       .catch(() => { });
   };
 
-  const handleDelete = (id: Entity.ID) => {
-    request(`/webtools/url-alias/delete/${id}`, { method: 'POST' })
+  const handleDelete = (id: string) => {
+    post(`/webtools/url-alias/delete/${id}`)
       .then(() => {
-        onDelete();
-        toggleNotification({ type: 'success', message: { id: 'webtools.settings.success.url_alias.delete' } });
+        if (onDelete) onDelete();
+        toggleNotification({ type: 'success', message: formatMessage({ id: 'webtools.settings.success.url_alias.delete' }) });
       })
       .catch(() => {
-        onDelete();
-        toggleNotification({ type: 'warning', message: { id: 'notification.error' } });
+        if (onDelete) onDelete();
+        toggleNotification({ type: 'warning', message: formatMessage({ id: 'notification.error' }) });
       });
   };
 
@@ -79,37 +76,35 @@ const TableRow: FC<Props> = ({
           {config.website_url && (
             <IconButton
               onClick={() => window.open(`${config.website_url}${row.url_path}`, '_blank')}
-              noBorder
-              icon={<ExternalLink />}
               label={formatMessage(
                 { id: 'webtools.settings.page.list.table.actions.edit', defaultMessage: 'Go to the front-end page' },
                 { target: `${row.url_path}` },
               )}
-            />
+            >
+              <ExternalLink />
+            </IconButton>
           )}
           <IconButton
             onClick={() => handleClick(row.url_path)}
-            noBorder
-            icon={<Pencil />}
             label={formatMessage(
               { id: 'webtools.settings.page.list.table.actions.goTo', defaultMessage: 'Edit {target}' },
               { target: `${row.url_path}` },
             )}
-          />
-          <IconButton
-            onClick={() => setOpenDeleteModal(true)}
-            noBorder
-            icon={<Trash />}
-            label={formatMessage(
-              { id: 'webtools.settings.page.list.table.actions.delete', defaultMessage: 'Delete {target}' },
-              { target: `${row.url_path}` },
-            )}
-          />
+          >
+            <Pencil />
+          </IconButton>
           <DeleteConfirmModal
-            isOpen={openDeleteModal}
-            onClose={() => setOpenDeleteModal(false)}
-            onSubmit={() => handleDelete(row.id)}
-          />
+            onSubmit={() => handleDelete(row.documentId)}
+          >
+            <IconButton
+              label={formatMessage(
+                { id: 'webtools.settings.page.list.table.actions.delete', defaultMessage: 'Delete {target}' },
+                { target: `${row.url_path}` },
+              )}
+            >
+              <Trash />
+            </IconButton>
+          </DeleteConfirmModal>
         </Flex>
       </Td>
     </Tr>
